refactor(foundation): split enums.rs into domain submodules (E7 #1)
enums.rs(4,919줄, 40 enums)를 enums/{paragraph,char_style,page,
border_fill,shape,tab,reference}.rs + mod.rs(re-export+테스트)로 분할.
각 enum/impl 본문 무수정 통째 이동. mod.rs 가 전부 pub use 재노출 →
hwpforge_foundation::enums::Foo 및 hwpforge_foundation::Foo 경로 불변.
106개 enum 테스트는 mod.rs 에 유지(use super::*).

순수 기계적 분할 — 동작·공개 API byte-동일. make ci 2536 통과.
